requirements.txt
django
django-ipware
ip2geotools
